put all visual tiles in one file and let hold a tile also a logical tile. Make tiles more easy to maintain
