+++ Frequency/Board/Map/Tile.py
@@ -1,10 +1,13 @@
+import GameLogic.Map
 from Helpers.EventHelpers import EventExist
 from Board.Units.Soldier import *
 from Board.Buildings.Barrack import Barracks
+from Vector2 import Vector2
+
 
 class Tile:
 
-    def __init__(self, position, defaultMoney, enemyMoney, texture, size, units=None, rectangle=None, building=None):
+    def __init__(self, position, defaultMoney, enemyMoney, texture, size, units, rectangle, building, logicTile):
         self.Position = position
         self.DefaultMoney = defaultMoney
         self.EnemyMoney = enemyMoney
@@ -12,11 +15,15 @@ class Tile:
         self.Texture = texture
         self.Rectangle = rectangle
         if units is None:
-           self.Units = []
+            self.Units = []
         else:
             self.Units = units
         self.Building = building
+        self._logicTile = logicTile
 
+    @property
+    def LogicTile(self) -> GameLogic.Map.Tile:
+        return self._logicTile
 
     def Update(self, game):
         if self.IsClickedByMouse(game):
@@ -28,8 +35,7 @@ class Tile:
                  if game.Logic.CanAddUnitBuildingToTile(game, self):
                     self.Building = Barracks(game.Logic.PlayingPlayer, self)
 
-        return Tile(self.Position, self.DefaultMoney, self.EnemyMoney, self.Texture, self.Size, self.Units, self.Rectangle, self.Building)
-
+        return type(self)(self.Position, self.Size, self.LogicTile, self.Texture, self.Units, self.Rectangle, self.Building)
 
     def Draw(self, game):
         screen = game.Settings.GetScreen()
@@ -42,10 +48,49 @@ class Tile:
         if self.Building is not None:
             self.Building.Draw(game)
 
-
     def IsHoverdByMouse(self):
         return self.Rectangle is not None and self.Rectangle.collidepoint(pygame.mouse.get_pos())
 
-
     def IsClickedByMouse(self, game):
         return self.IsHoverdByMouse() and EventExist(game.Events, pygame.MOUSEBUTTONUP)
+
+
+class DesertTile(Tile):
+
+    def __init__(self, position: Vector2, size: Vector2, logicTile, texture=None, units=None, rectangle=None, building=None):
+        texture = texture if texture is not None else pygame.transform.scale(pygame.image.load('images/tiles/DesertSeamless.png'), [size.X, size.Y])
+        super().__init__(position, 50, 100, texture, size, units, rectangle, building, logicTile)
+
+
+class GoldTile(Tile):
+
+    def __init__(self, position: Vector2, size: Vector2, logicTile, texture=None, units=None, rectangle=None, building=None):
+        texture = texture if texture is not None else pygame.transform.scale(pygame.image.load('images/tiles/GoldSeamless.png'), [size.X, size.Y])
+        super().__init__(position, 50, 100, texture, size, units, rectangle, building, logicTile)
+
+
+class ForestTile(Tile):
+
+    def __init__(self, position: Vector2, size: Vector2, logicTile, texture=None, units=None, rectangle=None, building=None):
+        texture = texture if texture is not None else pygame.transform.scale(pygame.image.load('images/tiles/ForestSeamless.png'), [size.X, size.Y])
+        super().__init__(position, 50, 100, texture, size, units, rectangle, building, logicTile)
+
+
+class IceTile(Tile):
+
+    def __init__(self, position: Vector2, size: Vector2, logicTile, texture=None, units=None, rectangle=None, building=None):
+        texture = texture if texture is not None else pygame.transform.scale(pygame.image.load('images/tiles/IceSeamless.png'), [size.X, size.Y])
+        super().__init__(position, 50, 100, texture, size, units, rectangle, building, logicTile)
+
+
+
+class SeaTile(Tile):
+    def __init__(self, position: Vector2, size: Vector2, logicTile, texture=None, units=None, rectangle=None, building=None):
+        texture = texture if texture is not None else pygame.transform.scale(pygame.image.load('images/tiles/SeaSeamless.png'), [size.X, size.Y])
+        super().__init__(position, 50, 100, texture, size, units, rectangle, building, logicTile)
+
+
+class SwampTile(Tile):
+    def __init__(self, position: Vector2, size: Vector2, logicTile, texture=None, units=None, rectangle=None, building=None):
+        texture = texture if texture is not None else pygame.transform.scale(pygame.image.load('images/tiles/SwampSeamless.png'), [size.X, size.Y])
+        super().__init__(position, 50, 100, texture, size, units, rectangle, building, logicTile)
